apps/infy_dpp_processor/requirements.txt
# [dev-packages]
pytest==6.2.5
pytest-cov==4.1.0
autopep8==2.0.4
# [packages]

python-dotenv==1.0.1
pipdeptree==2.*
./lib/infy_fs_utils-0.0.4-py3-none-any.whl[cloud]
./lib/infy_dpp_sdk-0.0.10-py3-none-any.whl
./lib/infy_gen_ai_sdk-0.0.6-py3-none-any.whl[st]
./lib/infy_common_utils-0.0.10-py3-none-any.whl
./lib/infy_ocr_generator-0.0.13-py3-none-any.whl[tesseract]
./lib/infy_ocr_parser-0.0.17-py3-none-any.whl
./lib/infy_dpp_core-0.0.3-py3-none-any.whl
./lib/infy_dpp_content_extractor-0.0.3-py3-none-any.whl
./lib/infy_dpp_segmentation-0.0.7-py3-none-any.whl
./lib/infy_dpp_ai-0.0.4-py3-none-any.whl
./lib/infy_dpp_storage-0.0.1-py3-none-any.whl
./lib/infy_table_extractor-0.0.12-py3-none-any.whl
